1/11/08 - fix misc stuff in PlayArea viewport management to handle MVT x-coordinate translation 1/11/08 - bail on PlayArea.handlePlayAreaResize if the canvases don't have non-zero sizes 1/11/08 - make MVT distance transforms require positive values

+++ trunk/simulations-java/simulations/glaciers/src/edu/colorado/phet/glaciers/module/basic/BasicModule.java
@@ -32,7 +32,7 @@ public class BasicModule extends PiccoloModule {
     
     private static final double VIEW_X_SCALE = 1.0;
     private static final double VIEW_Y_SCALE = 1.0;
-    private static final Point2D VIEW_ORIGIN = new Point2D.Double( 0, 0 );
+    private static final Point2D VIEW_ORIGIN = new Point2D.Double( 500, 0 );
     
     //----------------------------------------------------------------------------
     // Instance data
